Sweep collection happy path tests

+++ test/modules/CPortModule.t.sol
@@ -763,7 +763,7 @@ contract cPortModuleTest is Test, cPortEvents {
         _cPort.bulkBuyListingsCosigned{value: nativePaymentValue}(fnCalldata);
     }
 
-    function _sweepSignedListings(address caller, uint128 nativePaymentValue, SweepOrder memory sweepOrder, FuzzedOrder721[] memory fuzzedOrderInputsArray, Order[] memory saleDetailsArray, bytes4 expectedRevertSelector) internal {
+    function _sweepSignedListings(address caller, uint256 nativePaymentValue, SweepOrder memory sweepOrder, FuzzedOrder721[] memory fuzzedOrderInputsArray, Order[] memory saleDetailsArray, bytes4 expectedRevertSelector) internal {
         SignatureECDSA[] memory sellerSignaturesArray = new SignatureECDSA[](saleDetailsArray.length);
         SweepItem[] memory sweepItems = new SweepItem[](saleDetailsArray.length);
 
@@ -797,7 +797,7 @@ contract cPortModuleTest is Test, cPortEvents {
         _cPort.sweepCollection{value: nativePaymentValue}(fnCalldata);
     }
 
-    function _sweepSignedListingsWithFeeOnTop(address caller, uint128 nativePaymentValue, SweepOrder memory sweepOrder, FuzzedOrder721[] memory fuzzedOrderInputsArray, Order[] memory saleDetailsArray, FeeOnTop memory feeOnTop, bytes4 expectedRevertSelector) internal {
+    function _sweepSignedListingsWithFeeOnTop(address caller, uint256 nativePaymentValue, SweepOrder memory sweepOrder, FuzzedOrder721[] memory fuzzedOrderInputsArray, Order[] memory saleDetailsArray, FeeOnTop memory feeOnTop, bytes4 expectedRevertSelector) internal {
         SignatureECDSA[] memory sellerSignaturesArray = new SignatureECDSA[](saleDetailsArray.length);
         SweepItem[] memory sweepItems = new SweepItem[](saleDetailsArray.length);
 
@@ -836,7 +836,7 @@ contract cPortModuleTest is Test, cPortEvents {
         _cPort.sweepCollectionWithFeeOnTop{value: nativePaymentValue}(fnCalldata);
     }
 
-    function _sweepCosignedListings(address caller, uint128 nativePaymentValue, SweepOrder memory sweepOrder, FuzzedOrder721[] memory fuzzedOrderInputsArray, Order[] memory saleDetailsArray, bytes4 expectedRevertSelector) internal {
+    function _sweepCosignedListings(address caller, uint256 nativePaymentValue, SweepOrder memory sweepOrder, FuzzedOrder721[] memory fuzzedOrderInputsArray, Order[] memory saleDetailsArray, bytes4 expectedRevertSelector) internal {
         SignatureECDSA[] memory sellerSignaturesArray = new SignatureECDSA[](saleDetailsArray.length);
         Cosignature[] memory cosignaturesArray = new Cosignature[](saleDetailsArray.length);
         SweepItem[] memory sweepItems = new SweepItem[](saleDetailsArray.length);
@@ -872,7 +872,7 @@ contract cPortModuleTest is Test, cPortEvents {
         _cPort.sweepCollectionCosigned{value: nativePaymentValue}(fnCalldata);
     }
 
-    function _sweepEmptyCosignedListings(address caller, uint128 nativePaymentValue, SweepOrder memory sweepOrder, FuzzedOrder721[] memory fuzzedOrderInputsArray, Order[] memory saleDetailsArray, bytes4 expectedRevertSelector) internal {
+    function _sweepEmptyCosignedListings(address caller, uint256 nativePaymentValue, SweepOrder memory sweepOrder, FuzzedOrder721[] memory fuzzedOrderInputsArray, Order[] memory saleDetailsArray, bytes4 expectedRevertSelector) internal {
         SignatureECDSA[] memory sellerSignaturesArray = new SignatureECDSA[](saleDetailsArray.length);
         Cosignature[] memory cosignaturesArray = new Cosignature[](saleDetailsArray.length);
         SweepItem[] memory sweepItems = new SweepItem[](saleDetailsArray.length);
@@ -910,7 +910,7 @@ contract cPortModuleTest is Test, cPortEvents {
         _cPort.sweepCollectionCosigned{value: nativePaymentValue}(fnCalldata);
     }
 
-    function _sweepCosignedListingsWithFeeOnTop(address caller, uint128 nativePaymentValue, SweepOrder memory sweepOrder, FuzzedOrder721[] memory fuzzedOrderInputsArray, Order[] memory saleDetailsArray, FeeOnTop memory feeOnTop, bytes4 expectedRevertSelector) internal {
+    function _sweepCosignedListingsWithFeeOnTop(address caller, uint256 nativePaymentValue, SweepOrder memory sweepOrder, FuzzedOrder721[] memory fuzzedOrderInputsArray, Order[] memory saleDetailsArray, FeeOnTop memory feeOnTop, bytes4 expectedRevertSelector) internal {
         SignatureECDSA[] memory sellerSignaturesArray = new SignatureECDSA[](saleDetailsArray.length);
         Cosignature[] memory cosignaturesArray = new Cosignature[](saleDetailsArray.length);
         SweepItem[] memory sweepItems = new SweepItem[](saleDetailsArray.length);
@@ -951,7 +951,7 @@ contract cPortModuleTest is Test, cPortEvents {
         _cPort.sweepCollectionCosignedWithFeeOnTop{value: nativePaymentValue}(fnCalldata);
     }
 
-    function _sweepEmptyCosignedListingsWithFeeOnTop(address caller, uint128 nativePaymentValue, SweepOrder memory sweepOrder, FuzzedOrder721[] memory fuzzedOrderInputsArray, Order[] memory saleDetailsArray, FeeOnTop memory feeOnTop, bytes4 expectedRevertSelector) internal {
+    function _sweepEmptyCosignedListingsWithFeeOnTop(address caller, uint256 nativePaymentValue, SweepOrder memory sweepOrder, FuzzedOrder721[] memory fuzzedOrderInputsArray, Order[] memory saleDetailsArray, FeeOnTop memory feeOnTop, bytes4 expectedRevertSelector) internal {
         SignatureECDSA[] memory sellerSignaturesArray = new SignatureECDSA[](saleDetailsArray.length);
         Cosignature[] memory cosignaturesArray = new Cosignature[](saleDetailsArray.length);
         SweepItem[] memory sweepItems = new SweepItem[](saleDetailsArray.length);
